Add a return type using phpdoc to fix IDE autocomplete
The make function can't define a return type, because the result could be a Decorator, or the actual action.
However, without docblocks or type-hinting, PHPStorm ends up providing incorrect autocomplete information.
The solution is to instead add back a phpdoc comment for this specific use case.

See #285 and #287 for more details.